+++ RELEASENOTES.md
@@ -26,6 +26,8 @@
     *   Add `BasePreloadManager.Listener` to propagate preload events to apps.
     *   Allow changing SNTP client timeout
         ([#1540](https://github.com/androidx/media/issues/1540)).
+    *   Remove `MediaCodecAdapter.Configuration.flags` as the field was always
+        zero.
 *   Transformer:
     *   Add `SurfaceAssetLoader`, which supports queueing video data to
         Transformer via a `Surface`.